# Ocean/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Sóng nước của Đại Dương
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ShowcaseLink {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  likesCount: number;
  createdAt: string;
}

export interface ShowcaseLinkInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  /** @minLength 1 */
  url: string;
}

export interface LinksStats {
  totalLinks: number;
  totalLikes: number;
  latestLink: ShowcaseLink;
}