# Ocean/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Sóng nước của Đại Dương
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: links
    description: Showcase links operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /links:
    get:
      operationId: listLinks
      tags: [links]
      summary: List all showcase links
      responses:
        "200":
          description: List of showcase links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShowcaseLink"
    post:
      operationId: createLink
      tags: [links]
      summary: Create a new showcase link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShowcaseLinkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowcaseLink"
        "400":
          description: Invalid input

  /links/{id}:
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a showcase link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /links/{id}/like:
    post:
      operationId: likeLink
      tags: [links]
      summary: Increment likes count for a showcase link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated link with new likes count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowcaseLink"
        "404":
          description: Not found

  /links/stats:
    get:
      operationId: getLinksStats
      tags: [links]
      summary: Get aggregate stats for showcase links
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinksStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ShowcaseLink:
      type: object
      required: [id, title, url, likesCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        url:
          type: string
        likesCount:
          type: integer
        createdAt:
          type: string

    ShowcaseLinkInput:
      type: object
      required: [title, url]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        url:
          type: string
          minLength: 1

    LinksStats:
      type: object
      required: [totalLinks, totalLikes, latestLink]
      properties:
        totalLinks:
          type: integer
        totalLikes:
          type: integer
        latestLink:
          $ref: "#/components/schemas/ShowcaseLink"