# lib/api-zod/src/generated/types/showcaseLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Sóng nước của Đại Dương
 * OpenAPI spec version: 0.1.0
 */

export interface ShowcaseLink {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  url: string;
  likesCount: number;
  createdAt: string;
}